requirements.txt
# 核心依赖
torch==2.2.0
torchvision==0.17.0
albumentations==1.4.23
opencv-python-headless==4.10.0.84
numpy==1.24.4
matplotlib==3.7.1
tqdm==4.66.5
scikit-learn==1.3.0
Pillow==10.3.0

# 其他必要依赖
scipy==1.14.1
pandas==1.5.3